separate edit co-insured to not be imported by other modules
