+++ Projects/App/Sources/Journeys/CoInsuredJourney.swift
@@ -1,6 +1,7 @@
 import Contracts
 import EditCoInsured
 import Presentation
+import hCore
 
 extension AppJourney {
     @JourneyBuilder
