+++ Projects/App/Sources/Journeys/LoggedInJourney.swift
@@ -78,6 +78,18 @@ extension AppJourney {
                         let store: ContractStore = globalPresentableStoreContainer.get()
                         store.send(.fetch)
                     }
+            case let .handleCoInsured(config, fromInfoCard):
+                EditCoInsuredJourney.handleOpenEditCoInsured(for: config, fromInfoCard: fromInfoCard)
+                    .onDismiss {
+                        let store: ContractStore = globalPresentableStoreContainer.get()
+                        store.send(.fetch)
+                    }
+            case let .openMissingCoInsuredAlert(config):
+                EditCoInsuredJourney.openMissingCoInsuredAlert(config: config)
+                    .onDismiss {
+                        let store: ContractStore = globalPresentableStoreContainer.get()
+                        store.send(.fetch)
+                    }
             }
         }
         .makeTabSelected(UgglanStore.self) { action in
