+++ Projects/Contracts/Sources/Models/ContractModels.swift
@@ -1,4 +1,3 @@
-import EditCoInsured
 import Foundation
 import Presentation
 import hCore
