+++ Projects/Contracts/Sources/Stores/ContractState.swift
@@ -1,5 +1,4 @@
 import Apollo
-import EditCoInsured
 import Presentation
 import SwiftUI
 import hCore
