+++ Projects/EditCoInsured/Sources/Services/DemoImplementation/EditCoInsuredDemoClient.swift
@@ -1,4 +1,5 @@
 import Foundation
+import hCore
 import hGraphQL
 
 public class EditCoInsuredDemoClient: EditCoInsuredService {
