+++ Projects/EditCoInsured/Sources/Services/Protocols/EditCoInsuredService.swift
@@ -1,4 +1,5 @@
 import Foundation
+import hCore
 
 public protocol EditCoInsuredService {
     func sendMidtermChangeIntentCommit(commitId: String) async throws
