+++ Projects/EditCoInsured/Sources/Stores/EditCoInsuredAction.swift
@@ -1,4 +1,5 @@
 import Presentation
+import hCore
 
 public enum EditCoInsuredAction: ActionProtocol, Hashable {
     case openEditCoInsured(config: InsuredPeopleConfig, fromInfoCard: Bool)
