+++ Projects/EditCoInsured/Sources/View/AlreadyExistingCoInsured/InsuredPeopleScreen.swift
@@ -326,86 +326,3 @@ class InsuredPeopleNewScreenModel: ObservableObject {
         addCoInsured(coInsuredModel)
     }
 }
-
-public struct CoInsuredListType: Hashable, Identifiable {
-    public let id = UUID().uuidString
-    public init(
-        coInsured: CoInsuredModel,
-        type: StatusPillType? = nil,
-        date: String? = nil,
-        locallyAdded: Bool,
-        isContractOwner: Bool? = nil,
-        isEmpty: Bool? = false
-    ) {
-        self.coInsured = coInsured
-        self.type = type
-        self.date = date
-        self.locallyAdded = locallyAdded
-        self.isContractOwner = isContractOwner
-        self.isEmpty = isEmpty
-    }
-
-    public var coInsured: CoInsuredModel
-    public var type: StatusPillType?
-    public var date: String?
-    var locallyAdded: Bool
-    var isContractOwner: Bool?
-    public var isEmpty: Bool?
-}
-
-public struct InsuredPeopleConfig: Codable & Equatable & Hashable {
-    public var contractCoInsured: [CoInsuredModel]
-    public var contractId: String
-    public var activeFrom: String?
-    public var numberOfMissingCoInsured: Int
-    public var numberOfMissingCoInsuredWithoutTermination: Int
-    public let displayName: String
-    public let preSelectedCoInsuredList: [CoInsuredModel]
-    public let contractDisplayName: String
-    public let holderFirstName: String
-    public let holderLastName: String
-    public let holderSSN: String?
-    public var holderFullName: String {
-        return holderFirstName + " " + holderLastName
-    }
-
-    public init() {
-        self.contractCoInsured = []
-        self.contractId = ""
-        self.activeFrom = nil
-        self.numberOfMissingCoInsured = 0
-        self.numberOfMissingCoInsuredWithoutTermination = 0
-        self.displayName = ""
-        self.holderFirstName = ""
-        self.holderLastName = ""
-        self.holderSSN = nil
-        self.preSelectedCoInsuredList = []
-        self.contractDisplayName = ""
-    }
-
-    public init(
-        contractCoInsured: [CoInsuredModel],
-        contractId: String,
-        activeFrom: String?,
-        numberOfMissingCoInsured: Int,
-        numberOfMissingCoInsuredWithoutTermination: Int,
-        displayName: String,
-        preSelectedCoInsuredList: [CoInsuredModel],
-        contractDisplayName: String,
-        holderFirstName: String,
-        holderLastName: String,
-        holderSSN: String?
-    ) {
-        self.contractCoInsured = contractCoInsured
-        self.contractId = contractId
-        self.activeFrom = activeFrom
-        self.numberOfMissingCoInsured = numberOfMissingCoInsured
-        self.numberOfMissingCoInsuredWithoutTermination = numberOfMissingCoInsuredWithoutTermination
-        self.displayName = displayName
-        self.preSelectedCoInsuredList = preSelectedCoInsuredList
-        self.contractDisplayName = contractDisplayName
-        self.holderFirstName = holderFirstName
-        self.holderLastName = holderLastName
-        self.holderSSN = holderSSN
-    }
-}
