+++ Projects/Home/Sources/HomeState.swift
@@ -1,6 +1,5 @@
 import Apollo
 import Contracts
-import EditCoInsured
 import Foundation
 import Presentation
 import SwiftUI
