+++ Projects/Home/Sources/Screens/CoInsuredInfoHomeView.swift
@@ -1,0 +1,25 @@
+import SwiftUI
+import hCore
+import hCoreUI
+
+public struct CoInsuredInfoHomeView: View {
+    var onTapAction: () -> Void
+
+    public init(
+        onTapAction: @escaping () -> Void
+    ) {
+        self.onTapAction = onTapAction
+    }
+
+    public var body: some View {
+        InfoCard(text: L10n.contractCoinsuredMissingInfoText, type: .attention)
+            .buttons([
+                .init(
+                    buttonTitle: L10n.contractCoinsuredMissingAddInfo,
+                    buttonAction: {
+                        onTapAction()
+                    }
+                )
+            ])
+    }
+}
