+++ Projects/Home/Sources/Screens/Home.swift
@@ -1,7 +1,6 @@
 import Apollo
 import Combine
 import Contracts
-import EditCoInsured
 import Foundation
 import Payment
 import Presentation
