+++ Projects/TravelCertificate/Sources/Views/WhoIsTravelingScreen.swift
@@ -1,5 +1,4 @@
 import Contracts
-import EditCoInsured
 import Foundation
 import Presentation
 import SwiftUI
